requirements.txt
fastapi[standard]==0.116.2
uvicorn==0.35.0
torch==2.8.0
sentence-transformers==5.1.1
loguru==0.7.3
einops==0.8.1
pydantic-settings==2.11.0
pyyaml==6.0.3
ruff==0.14.3